fix: accept fractional SchaleDB skill values (#56)
## Summary

- Preserve fractional skill effect values instead of skipping affected
students during SchaleDB import.
- Add a regression test using the fractional EX effect shape from
student 20060.

## Verification

Static checks: PASS (changed files: 2).
The Go test suite verifies successful decoding and exact preservation of
the fractional value.

+++ internal/types/schaledb_students.go
@@ -130,7 +130,7 @@ type SkillEffect struct {
 	Restrictions         []SkillRestriction    `json:"Restrictions,omitempty"`
 	Target               []string              `json:"Target,omitempty"`
 	Scale                []int                 `json:"Scale,omitempty"`
-	Value                [][]int               `json:"Value,omitempty"`
+	Value                [][]float64           `json:"Value,omitempty"`
 	AdditionalValue      any                   `json:"AdditionalValue,omitempty"`
 	Stat                 string                `json:"Stat,omitempty"`
 	Duration             int                   `json:"Duration,omitempty"`
